fix(serving): BigInteger model sizes + honest, non-blocking deploy health
Two serving fixes surfaced while live-verifying the Deploy tab.

- size_bytes was SQLAlchemy Integer (Postgres INTEGER caps at ~2.147 GB), so
  seeding any GGUF >= ~2.1 GB (a 7B Q4 is ~4 GB) overflowed on insert and wedged
  the seed job in a retry loop. Widen to BigInteger. create_all has no migration
  framework, so existing DBs need a manual
  `ALTER TABLE model_store_entry ALTER COLUMN size_bytes TYPE BIGINT`.
- A deployment's health was frozen at "starting / Connection refused": deploy
  spawned llama-server and health-probed it once, before the model finished
  loading, and nothing re-probed. Add PersistentSupervisor.await_ready(), which
  re-runs reconcile() (the probe runs in the worker, where 127.0.0.1:<port> is
  reachable) until the record is healthy or a bounded timeout, so the Deploy tab
  shows a truthful running state. Raise that deploy spec's health_failure_threshold
  so the readiness poll can't trip reconcile's degrade-and-kill mid-load.
- await_ready blocks (bounded time.sleep), so ControlPlane.up now runs the sync
  serve_store_model via asyncio.to_thread — otherwise it would stall the scale-1
  worker's asyncio loop (Inngest Connect heartbeats + concurrent steps) while a
  large GGUF loads, risking a heartbeat-miss redeliver and a double-deploy port
  conflict.

+++ small-doc-ie-bench/src/docie_bench/serving/catalog.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 import datetime as dt
 from typing import Any
 
-from sqlalchemy import DateTime, Integer, String, Text, select
+from sqlalchemy import BigInteger, DateTime, String, Text, select
 from sqlalchemy.orm import Mapped, mapped_column
 
 from docie_bench.serving.model_store import FAMILIES, StoreEntry
@@ -35,7 +35,11 @@ class ModelStoreEntry(Base):
     model_path: Mapped[str] = mapped_column(Text)
     mmproj_path: Mapped[str | None] = mapped_column(Text, nullable=True)
     source: Mapped[str | None] = mapped_column(String(512), nullable=True)
-    size_bytes: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    # BigInteger: GGUF blobs routinely exceed Postgres INTEGER's ~2.147 GB cap
+    # (a 7B Q4 is ~4 GB), which would overflow on insert. This app uses
+    # create_all (no migrations), so an EXISTING database must be migrated
+    # manually: ALTER TABLE model_store_entry ALTER COLUMN size_bytes TYPE BIGINT.
+    size_bytes: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     created_at: Mapped[dt.datetime] = mapped_column(DateTime(timezone=True), default=_utcnow)
     updated_at: Mapped[dt.datetime] = mapped_column(
         DateTime(timezone=True), default=_utcnow, onupdate=_utcnow
